Automatically add video dimensions Fixes #325
