+++ package.json
@@ -11,7 +11,7 @@
     "sw": "workbox injectManifest workbox-config.js && terser dist/sw.js -cm -o dist/sw.js",
     "copy-img": "shx cp ./node_modules/@browser-logos/{chrome,firefox}/*.svg src/_css/img/ && shx cp ./node_modules/@browser-logos/safari/safari_256x256.png src/_css/img/safari.png",
     "copy-js": "shx cp ./node_modules/dark-mode-toggle/dist/dark-mode-toggle.min.mjs dist/_js/dark-mode-toggle.mjs",
-    "img": "node add-image-dimensions.js",
+    "img-video-dimensions": "node add-image-and-video-dimensions.js",
     "clean": "shx rm -rf -- dist",
     "build": "npm run clean && npm run css && npm run js && npm run json && npm run copy-img && npm run copy-js && npm run eleventy && npm run sw && npm run xml && npm run html",
     "watch": "npm run build && npx eleventy --watch",
@@ -42,6 +42,7 @@
     "cssnano": "^4.1.10",
     "dark-mode-toggle": "^0.5.1",
     "firebase-tools": "^7.2.4",
+    "get-video-dimensions": "^1.0.0",
     "glob": "^7.1.4",
     "he": "^1.2.0",
     "html-minifier": "^4.0.0",
